Test user dir rebuilding in storage, not handlers
To do this:
 - move required helpers to the handlers test dir and pull them in via
 inheritance (ew)

Additionally, we no longer rebuild the user dir the the user dir
handlers tests. Instead, tests register their own users and that creates
the required user_dir entries.
